Factor out test_utils.
Necessary to prevent accidential inclusion of test flags.
