+++ clients/csv/exporter/exporter_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/attic-labs/noms/chunks"
-	"github.com/attic-labs/noms/clients/util"
+	"github.com/attic-labs/noms/clients/test_util"
 	"github.com/attic-labs/noms/d"
 	"github.com/attic-labs/noms/datas"
 	"github.com/attic-labs/noms/dataset"
@@ -21,7 +21,7 @@ func TestCSVExporter(t *testing.T) {
 }
 
 type testSuite struct {
-	util.ClientTestSuite
+	test_util.ClientTestSuite
 }
 
 // FIXME: run with pipe
