+++ clients/test_util/client_test_suite.go
@@ -1,4 +1,4 @@
-package util
+package test_util
 
 import (
 	"io/ioutil"
