+++ clients/test_util/util_test.go
@@ -1,8 +1,9 @@
-package util
+package test_util
 
 import (
 	"testing"
 
+	"github.com/attic-labs/noms/clients/util"
 	"github.com/attic-labs/noms/types"
 	"github.com/stretchr/testify/suite"
 )
@@ -16,38 +17,38 @@ type LibTestSuite struct {
 }
 
 func (suite *LibTestSuite) TestPrimitiveTypes() {
-	suite.EqualValues(types.NewString("expected"), NomsValueFromDecodedJSON("expected"))
-	suite.EqualValues(types.Bool(false), NomsValueFromDecodedJSON(false))
-	suite.EqualValues(types.Number(1.7), NomsValueFromDecodedJSON(1.7))
-	suite.False(NomsValueFromDecodedJSON(1.7).Equals(types.Bool(true)))
+	suite.EqualValues(types.NewString("expected"), util.NomsValueFromDecodedJSON("expected"))
+	suite.EqualValues(types.Bool(false), util.NomsValueFromDecodedJSON(false))
+	suite.EqualValues(types.Number(1.7), util.NomsValueFromDecodedJSON(1.7))
+	suite.False(util.NomsValueFromDecodedJSON(1.7).Equals(types.Bool(true)))
 }
 
 func (suite *LibTestSuite) TestCompositeTypes() {
 	// [false true]
 	suite.EqualValues(
 		types.NewList().Append(types.Bool(false)).Append(types.Bool(true)),
-		NomsValueFromDecodedJSON([]interface{}{false, true}))
+		util.NomsValueFromDecodedJSON([]interface{}{false, true}))
 
 	// [[false true]]
 	suite.EqualValues(
 		types.NewList().Append(
 			types.NewList().Append(types.Bool(false)).Append(types.Bool(true))),
-		NomsValueFromDecodedJSON([]interface{}{[]interface{}{false, true}}))
+		util.NomsValueFromDecodedJSON([]interface{}{[]interface{}{false, true}}))
 
 	// {"string": "string",
 	//  "list": [false true],
 	//  "map": {"nested": "string"}
 	// }
-	m := NewMapOfStringToValue(
+	m := util.NewMapOfStringToValue(
 		types.NewString("string"),
 		types.NewString("string"),
 		types.NewString("list"),
 		types.NewList().Append(types.Bool(false)).Append(types.Bool(true)),
 		types.NewString("map"),
-		NewMapOfStringToValue(
+		util.NewMapOfStringToValue(
 			types.NewString("nested"),
 			types.NewString("string")))
-	o := NomsValueFromDecodedJSON(map[string]interface{}{
+	o := util.NomsValueFromDecodedJSON(map[string]interface{}{
 		"string": "string",
 		"list":   []interface{}{false, true},
 		"map":    map[string]interface{}{"nested": "string"},
@@ -57,5 +58,5 @@ func (suite *LibTestSuite) TestCompositeTypes() {
 }
 
 func (suite *LibTestSuite) TestPanicOnUnsupportedType() {
-	suite.Panics(func() { NomsValueFromDecodedJSON(map[int]string{1: "one"}) }, "Should panic on map[int]string!")
+	suite.Panics(func() { util.NomsValueFromDecodedJSON(map[int]string{1: "one"}) }, "Should panic on map[int]string!")
 }
